# clients/go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: email@gmail.com
  description: |
    This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Open APi
  version: 1.0.0
servers:
- url: http://localhost:8080
...
paths:
  /comment:
    post:
      operationId: addComment
      requestBody:
        $ref: '#/components/requestBodies/Comment'
      responses:
        "200":
          description: Comment created successfully
        "405":
          description: Invalid input
      security:
      - cm_auth:
        - write
        - read
      summary: Add a new comment
      tags:
      - comment
    put:
      operationId: updateComment
      requestBody:
        $ref: '#/components/requestBodies/Comment'
      responses:
        "200":
          description: Comment updated successfully
        "400":
          description: Invalid ID supplied
        "404":
          description: Comment not found
        "405":
          description: Validation exception
      security:
      - cm_auth:
        - write
        - read
      summary: Update an existing comment
      tags:
      - comment
  /comment/{commentId}:
    delete:
      operationId: deleteComment
      parameters:
      - explode: false
        in: header
        name: api_key
        required: false
        schema:
          type: string
        style: simple
      - description: Comment id to delete
        explode: false
        in: path
        name: commentId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Comment deleted successfully
        "400":
          description: Invalid ID supplied
        "404":
          description: Comment not found
      security:
      - cm_auth:
        - write
        - read
      summary: Deletes a comment
      tags:
      - comment
    get:
      description: Returns a single comment
      operationId: getCommentById
      parameters:
      - description: ID of comment to return
        explode: false
        in: path
        name: commentId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
            application/xml:
              schema:
                $ref: '#/components/schemas/Comment'
          description: Comment fetched successfully
        "400":
          description: Invalid ID supplied
        "404":
          description: Comment not found
      security:
      - api_key: []
      summary: Find comment by ID
      tags:
      - comment
    post:
      operationId: updateCommentWithForm
      parameters:
      - description: ID of comment that needs to be updated
        explode: false
        in: path
        name: commentId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/updateCommentWithForm_request'
      responses:
        "200":
          description: Comment created successfully
        "405":
          description: Invalid input
      security:
      - cm_auth:
        - write
        - read
      summary: Updates a comment with form data
      tags:
      - comment
...
  /blog:
    post:
      operationId: addBlog
      requestBody:
        $ref: '#/components/requestBodies/Blog'
      responses:
        "200":
          description: Blog created successfully
        "405":
          description: Invalid input
      security:
      - cm_auth:
        - write
        - read
      summary: Add a new blog
      tags:
      - blog
    put:
      operationId: updateBlog
      requestBody:
        $ref: '#/components/requestBodies/Blog'
      responses:
        "200":
          description: Blog updated successfully
        "400":
          description: Invalid ID supplied
        "404":
          description: Blog not found
        "405":
          description: Validation exception
      security:
      - cm_auth:
        - write
        - read
      summary: Update an existing blog
      tags:
      - blog
  /blog/{blogId}:
    delete:
      operationId: deleteBlog
      parameters:
      - explode: false
        in: header
        name: api_key
        required: false
        schema:
          type: string
        style: simple
      - description: Blog id to delete
        explode: false
        in: path
        name: blogId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Blog deleted successfully
        "400":
          description: Invalid ID supplied
        "404":
          description: Blog not found
      security:
      - cm_auth:
        - write
        - read
      summary: Deletes a blog
      tags:
      - blog
    get:
      description: Returns a single blog
      operationId: getBlogById
      parameters:
      - description: ID of blog to return
        explode: false
        in: path
        name: blogId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
            application/xml:
              schema:
                $ref: '#/components/schemas/Blog'
          description: blog fetched successfully
        "400":
          description: Invalid ID supplied
        "404":
          description: Blog not found
      security:
      - api_key: []
      summary: Find blog by ID
      tags:
      - blog
    post:
      operationId: updateBlogWithForm
      parameters:
      - description: ID of blog that needs to be updated
        explode: false
        in: path
        name: blogId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/updateBlogWithForm_request'
      responses:
        "200":
          description: Blog updated successfully
        "405":
          description: Invalid input
      security:
      - cm_auth:
        - write
        - read
      summary: Updates a blog with form data
      tags:
      - blog
...
components:
  requestBodies:
    Comment:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Comment'
        application/xml:
          schema:
            $ref: '#/components/schemas/Comment'
      description: Comment object that needs to be added
      required: true
...
    Blog:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Blog'
        application/xml:
          schema:
            $ref: '#/components/schemas/Blog'
      description: Blog object that needs to be added
  schemas:
    Comment:
      example:
        created_timestamp: 2000-01-23T04:56:07.000+00:00
        postId: postId
        media:
          name: name
          id: id
          type: type
          content: content
        authorDisplayName: authorDisplayName
        userId: userId
        parentId: parentId
        content: content
        modified_timestamp: 2000-01-23T04:56:07.000+00:00
        upvoteCount: 6
        downvoteCount: 0
        id: id
        blogId: blogId
        status: Draft
      properties:
        id:
          type: string
        userId:
          type: string
        postId:
          type: string
        blogId:
          type: string
        parentId:
          type: string
        authorDisplayName:
          type: string
        created_timestamp:
          format: date-time
          type: string
        modified_timestamp:
          format: date-time
          type: string
        content:
          type: string
        status:
          enum:
          - Draft
          - InReview
          - Flagged
          - Published
          type: string
        downvoteCount:
          type: integer
        upvoteCount:
          type: integer
        media:
          $ref: '#/components/schemas/Media'
      type: object
      xml:
        name: Comment
    Media:
      example:
        name: name
        id: id
        type: type
        content: content
      properties:
        id:
          type: string
        type:
          type: string
        name:
          type: string
        content:
          type: string
      type: object
      xml:
        name: Media
...
    Blog:
      example:
        authToken: authToken
        name: name
        id: id
        url: url
        commentStatus: Disabled
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        authToken:
          type: string
        commentStatus:
          enum:
          - Disabled
          - RequireReview
          - AutomateReview
          type: string
      type: object
      xml:
        name: Blog
    updateCommentWithForm_request:
      properties:
        userId:
          type: string
        postId:
          type: string
        blogId:
          type: string
        parentId:
          type: string
        authorDisplayName:
          type: string
        created_timestamp:
          format: date-time
          type: string
        modified_timestamp:
          format: date-time
          type: string
        content:
          type: string
        status:
          enum:
          - Draft
          - InReview
          - Flagged
          - Published
          type: string
        downvoteCount:
          type: integer
        upvoteCount:
          type: integer
        media:
          $ref: '#/components/schemas/Media'
      type: object
...
    updateBlogWithForm_request:
      properties:
        name:
          type: string
        url:
          type: string
        authToken:
          type: string
        commentStatus:
          enum:
          - Disabled
          - RequireReview
          - AutomateReview
          type: string
      type: object
  securitySchemes:
    cm_auth:
      flows:
        implicit:
          authorizationUrl: http://petstore.swagger.io/oauth/dialog
          scopes:
            write: modify
            read: read
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey