# servers/go/go/model_blog.go
/*
 * Open APi
 *
 * This is a sample server. 
 *
 * API version: 1.0.0
 * Contact: email@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Blog struct {

	Id string `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	Url string `json:"url,omitempty"`

	AuthToken string `json:"authToken,omitempty"`

	CommentStatus string `json:"commentStatus,omitempty"`
}

// AssertBlogRequired checks if the required fields are not zero-ed
func AssertBlogRequired(obj Blog) error {
	return nil
}

// AssertRecurseBlogRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of Blog (e.g. [][]Blog), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseBlogRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aBlog, ok := obj.(Blog)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertBlogRequired(aBlog)
	})
}