# servers/go/go/model_media.go
/*
 * Open APi
 *
 * This is a sample server. 
 *
 * API version: 1.0.0
 * Contact: email@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Media struct {

	Id string `json:"id,omitempty"`

	Type string `json:"type,omitempty"`

	Name string `json:"name,omitempty"`

	Content string `json:"content,omitempty"`
}

// AssertMediaRequired checks if the required fields are not zero-ed
func AssertMediaRequired(obj Media) error {
	return nil
}

// AssertRecurseMediaRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of Media (e.g. [][]Media), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseMediaRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aMedia, ok := obj.(Media)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertMediaRequired(aMedia)
	})
}